CT-8 DatatableSeeder refactoring
