+++ CodingTracker.Wolfieeex/CodingTracker.Wolfieeex/appsettings.json
@@ -13,7 +13,6 @@
   },
 
   "MockDatabaseOptions": {
-    "First": "F1",
     "BaseNumberOfLines": 400,
     "BaseMinYear": 2023,
     "BaseMinTime": "00:45:00",
